llm/simple_openai/requirements.txt
openai==1.30.1
python-dotenv==1.0.1